abstractmacros => ts and move out some utils
